refact(common/stats-charts): removing duplicates and providing a series of utilities for building charts
