+++ dashboard/src/common/stats-charts/components/select-date-view.tsx
