+++ dashboard/src/common/stats-charts/types.ts
@@ -1,0 +1,2 @@
+
+export type ChartDateInterval = '90d' | '30d' | '7d' | '1d';
