+++ dashboard/src/features/total-traffic-widget/components/index.ts
@@ -1,2 +1,1 @@
 export * from "./skeleton"
-export * from "./select-date-view"
