+++ dashboard/src/features/total-traffic-widget/types.ts
@@ -8,5 +8,3 @@ export interface NodesUsage {
 export interface UserNodesUsageWidgetProps {
     user: UserType;
 }
-
-export type ChartDateInterval = '90d' | '30d' | '7d' | '1d';
